// src/UI/components/dataset/dataset.action.types.ts
export enum DatasetActionTypeEnum {
  ASSIGN_PRIMARY_REVIEWER = 'Assign Primary Reviewer',
  ASSIGN_TERTIARY_REVIEWER = 'Assign Tertiary Reviewer',
  UPLOAD_PRIMARY_REVIEWED_DATASET = 'Upload Primary Reviewed Dataset',
  UPLOAD_TERTIARY_REVIEWED_DATASET = 'Upload Tertiary Reviewed Dataset',
  REJECT = 'Reject Dataset',
  APPROVE = 'Approve Dataset',
  SEND_EMAIL = 'Send Email',
}